+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintPlugin.kt
@@ -11,6 +11,8 @@ import org.gradle.api.internal.HasConvention
 import org.gradle.api.logging.configuration.ConsoleOutput
 import org.gradle.api.plugins.Convention
 import org.gradle.api.plugins.JavaPluginConvention
+import org.gradle.api.tasks.TaskProvider
+import org.gradle.language.base.plugins.LifecycleBasePlugin
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 import org.jetbrains.kotlin.gradle.plugin.KotlinPlatformType
 import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
@@ -24,36 +26,34 @@ import kotlin.reflect.KClass
 /**
  * Plugin that provides a wrapper over the `ktlint` project.
  */
+@Suppress("UnstableApiUsage")
 open class KtlintPlugin : Plugin<Project> {
 
     override fun apply(target: Project) {
-        val extension = target.plugins.apply(KtlintBasePlugin::class.java).extension
+        val holder = PluginHolder(target)
         // Apply the idea plugin
         target.plugins.apply(KtlintIdeaPlugin::class.java)
 
-        addKtLintTasksToKotlinPlugin(target, extension)
+        holder.addKtLintTasksToKotlinPlugin()
     }
 
-    private fun addKtLintTasksToKotlinPlugin(target: Project, extension: KtlintExtension) {
-        target.plugins.withId("kotlin", applyKtLint(target, extension))
-        target.plugins.withId("kotlin2js", applyKtLint(target, extension))
-        target.plugins.withId("kotlin-platform-common", applyKtLint(target, extension))
-        target.plugins.withId("kotlin-android", applyKtLintToAndroid(target, extension))
-        target.plugins.withId("konan", applyKtLintKonanNative(target, extension))
+    private fun PluginHolder.addKtLintTasksToKotlinPlugin() {
+        target.plugins.withId("kotlin", applyKtLint())
+        target.plugins.withId("kotlin2js", applyKtLint())
+        target.plugins.withId("kotlin-platform-common", applyKtLint())
+        target.plugins.withId("kotlin-android", applyKtLintToAndroid())
+        target.plugins.withId("konan", applyKtLintKonanNative())
         target.plugins.withId(
             "org.jetbrains.kotlin.native",
-            applyKtLintNative(target, extension)
+            applyKtLintNative()
         )
         target.plugins.withId(
             "org.jetbrains.kotlin.multiplatform",
-            applyKtlintMultiplatform(target, extension)
+            applyKtlintMultiplatform()
         )
     }
 
-    private fun applyKtlintMultiplatform(
-        target: Project,
-        extension: KtlintExtension
-    ): (Plugin<in Any>) -> Unit = {
+    private fun PluginHolder.applyKtlintMultiplatform(): (Plugin<in Any>) -> Unit = {
         val ktLintConfig = createConfiguration(target, extension)
         val multiplatformExtension = target.extensions.getByType(KotlinMultiplatformExtension::class.java)
 
@@ -66,7 +66,7 @@ open class KtlintPlugin : Plugin<Project> {
                 sourceSet.kotlin.sourceDirectories
             )
 
-            addKtlintCheckTaskToProjectMetaCheckTask(target, checkTask)
+            addKtlintCheckTaskToProjectMetaCheckTask(checkTask)
             setCheckTaskDependsOnKtlintCheckTask(target, checkTask)
 
             val ktlintSourceSetFormatTask = createFormatTask(
@@ -77,7 +77,7 @@ open class KtlintPlugin : Plugin<Project> {
                 sourceSet.kotlin.sourceDirectories
             )
 
-            addKtlintFormatTaskToProjectMetaFormatTask(target, ktlintSourceSetFormatTask)
+            addKtlintFormatTaskToProjectMetaFormatTask(ktlintSourceSetFormatTask)
         }
 
         multiplatformExtension.targets.all { kotlinTarget ->
@@ -99,14 +99,13 @@ open class KtlintPlugin : Plugin<Project> {
         }
     }
 
-    private fun applyKtLint(
-        target: Project,
-        extension: KtlintExtension
-    ): (Plugin<in Any>) -> Unit {
+    private fun PluginHolder.applyKtLint(): (Plugin<in Any>) -> Unit {
         return {
             val ktLintConfig = createConfiguration(target, extension)
 
-            target.theHelper<JavaPluginConvention>().sourceSets.forEach { sourceSet ->
+            val sourceSets = target.theHelper<JavaPluginConvention>().sourceSets
+
+            sourceSets.all { sourceSet ->
                 val kotlinSourceSet: SourceDirectorySet = (sourceSet as HasConvention)
                     .convention
                     .getPluginHelper<KotlinSourceSet>()
@@ -119,7 +118,7 @@ open class KtlintPlugin : Plugin<Project> {
                     kotlinSourceSet.sourceDirectories
                 )
 
-                addKtlintCheckTaskToProjectMetaCheckTask(target, checkTask)
+                addKtlintCheckTaskToProjectMetaCheckTask(checkTask)
                 setCheckTaskDependsOnKtlintCheckTask(target, checkTask)
 
                 val ktlintSourceSetFormatTask = createFormatTask(
@@ -130,15 +129,12 @@ open class KtlintPlugin : Plugin<Project> {
                     kotlinSourceSet.sourceDirectories
                 )
 
-                addKtlintFormatTaskToProjectMetaFormatTask(target, ktlintSourceSetFormatTask)
+                addKtlintFormatTaskToProjectMetaFormatTask(ktlintSourceSetFormatTask)
             }
         }
     }
 
-    private fun applyKtLintToAndroid(
-        target: Project,
-        extension: KtlintExtension
-    ): (Plugin<in Any>) -> Unit {
+    private fun PluginHolder.applyKtLintToAndroid(): (Plugin<in Any>) -> Unit {
         return {
             val ktLintConfig = createConfiguration(target, extension)
 
@@ -154,7 +150,7 @@ open class KtlintPlugin : Plugin<Project> {
                     sources
                 )
 
-                addKtlintCheckTaskToProjectMetaCheckTask(target, checkTask)
+                addKtlintCheckTaskToProjectMetaCheckTask(checkTask)
                 setCheckTaskDependsOnKtlintCheckTask(target, checkTask)
 
                 val ktlintSourceSetFormatTask = createFormatTask(
@@ -165,7 +161,7 @@ open class KtlintPlugin : Plugin<Project> {
                     sources
                 )
 
-                addKtlintFormatTaskToProjectMetaFormatTask(target, ktlintSourceSetFormatTask)
+                addKtlintFormatTaskToProjectMetaFormatTask(ktlintSourceSetFormatTask)
             }
 
             /*
@@ -206,26 +202,29 @@ open class KtlintPlugin : Plugin<Project> {
         multiplatformTargetName: String? = null
     ) {
         variants?.all { variant ->
-            val variantCheckTask = target.getAndroidVariantMetaKtlintCheckTask(variant.name, multiplatformTargetName)
-            val variantFormatTask = target.getAndroidVariantMetaKtlintFormatTask(variant.name, multiplatformTargetName)
+            val variantCheckTask = target.createAndroidVariantMetaKtlintCheckTask(
+                variant.name,
+                multiplatformTargetName
+            )
+            val variantFormatTask = target.createAndroidVariantMetaKtlintFormatTask(
+                variant.name,
+                multiplatformTargetName
+            )
             variant.sourceSets.forEach { sourceSet ->
                 val sourceSetName = "${multiplatformTargetName?.capitalize() ?: ""}${sourceSet.name.capitalize()}"
-                variantCheckTask.dependsOn(sourceSetName.sourceSetCheckTaskName())
-                variantFormatTask.dependsOn(sourceSetName.sourceSetFormatTaskName())
+                variantCheckTask.configure { it.dependsOn(sourceSetName.sourceSetCheckTaskName()) }
+                variantFormatTask.configure { it.dependsOn(sourceSetName.sourceSetFormatTaskName()) }
             }
         }
     }
 
-    private fun applyKtLintKonanNative(
-        project: Project,
-        extension: KtlintExtension
-    ): (Plugin<in Any>) -> Unit {
+    private fun PluginHolder.applyKtLintKonanNative(): (Plugin<in Any>) -> Unit {
         return {
-            val ktLintConfig = createConfiguration(project, extension)
+            val ktLintConfig = createConfiguration(target, extension)
 
-            val compileTargets = project.theHelper<KonanExtension>().targets
-            project.theHelper<KonanArtifactContainer>().whenObjectAdded { buildConfig ->
-                addTasksForNativePlugin(project, extension, buildConfig.name, ktLintConfig) {
+            val compileTargets = target.theHelper<KonanExtension>().targets
+            target.theHelper<KonanArtifactContainer>().whenObjectAdded { buildConfig ->
+                addTasksForNativePlugin(buildConfig.name, ktLintConfig) {
                     compileTargets.fold(initial = emptyList()) { acc, target ->
                         val compileTask = buildConfig.findByTarget(target)
                         if (compileTask != null) {
@@ -240,15 +239,11 @@ open class KtlintPlugin : Plugin<Project> {
         }
     }
 
-    private fun applyKtLintNative(
-        project: Project,
-        extension: KtlintExtension
-    ): (Plugin<in Any>) -> Unit {
+    private fun PluginHolder.applyKtLintNative(): (Plugin<in Any>) -> Unit {
         return {
-            val ktLintConfig = createConfiguration(project, extension)
-
-            project.components.withType(KotlinNativeComponent::class.java) { component ->
-                addTasksForNativePlugin(project, extension, component.name, ktLintConfig) {
+            val ktLintConfig = createConfiguration(target, extension)
+            target.components.withType(KotlinNativeComponent::class.java) { component ->
+                addTasksForNativePlugin(component.name, ktLintConfig) {
                     component.konanTargets.get()
                         .fold(initial = emptyList()) { acc, nativeTarget ->
                             acc + listOf(component.sources.getAllSources(nativeTarget))
@@ -258,9 +253,7 @@ open class KtlintPlugin : Plugin<Project> {
         }
     }
 
-    private fun addTasksForNativePlugin(
-        project: Project,
-        extension: KtlintExtension,
+    private fun PluginHolder.addTasksForNativePlugin(
         sourceSetName: String,
         ktlintConfiguration: Configuration,
         gatherVariantSources: () -> List<FileCollection>
@@ -268,38 +261,36 @@ open class KtlintPlugin : Plugin<Project> {
         val sourceDirectoriesList = gatherVariantSources()
         if (sourceDirectoriesList.isNotEmpty()) {
             val checkTask = createCheckTask(
-                project,
+                target,
                 extension,
                 sourceSetName,
                 ktlintConfiguration,
                 sourceDirectoriesList
             )
-            addKtlintCheckTaskToProjectMetaCheckTask(project, checkTask)
-            setCheckTaskDependsOnKtlintCheckTask(project, checkTask)
+            addKtlintCheckTaskToProjectMetaCheckTask(checkTask)
+            setCheckTaskDependsOnKtlintCheckTask(target, checkTask)
 
             val ktlintSourceSetFormatTask = createFormatTask(
-                project,
+                target,
                 extension,
                 sourceSetName,
                 ktlintConfiguration,
                 sourceDirectoriesList
             )
-            addKtlintFormatTaskToProjectMetaFormatTask(project, ktlintSourceSetFormatTask)
+            addKtlintFormatTaskToProjectMetaFormatTask(ktlintSourceSetFormatTask)
         }
     }
 
-    private fun addKtlintCheckTaskToProjectMetaCheckTask(target: Project, checkTask: Task) {
-        target.getMetaKtlintCheckTask().dependsOn(checkTask)
-        if (target.rootProject != target) {
-            target.rootProject.getMetaKtlintCheckTask().dependsOn(checkTask)
-        }
+    private fun PluginHolder.addKtlintCheckTaskToProjectMetaCheckTask(
+        checkTask: TaskProvider<KtlintCheckTask>
+    ) {
+        metaKtlintCheckTask.configure { it.dependsOn(checkTask) }
     }
 
-    private fun addKtlintFormatTaskToProjectMetaFormatTask(target: Project, formatTask: Task) {
-        target.getMetaKtlintFormatTask().dependsOn(formatTask)
-        if (target.rootProject != target) {
-            target.rootProject.getMetaKtlintFormatTask().dependsOn(formatTask)
-        }
+    private fun PluginHolder.addKtlintFormatTaskToProjectMetaFormatTask(
+        formatTask: TaskProvider<KtlintFormatTask>
+    ) {
+        metaKtlintFormatTask.configure { it.dependsOn(formatTask) }
     }
 
     private fun createFormatTask(
@@ -308,8 +299,8 @@ open class KtlintPlugin : Plugin<Project> {
         sourceSetName: String,
         ktLintConfig: Configuration,
         kotlinSourceDirectories: Iterable<*>
-    ): Task {
-        return target.taskHelper<KtlintFormatTask>(sourceSetName.sourceSetFormatTaskName()) {
+    ): TaskProvider<KtlintFormatTask> {
+        return target.registerTask(sourceSetName.sourceSetFormatTaskName()) {
             description = "Runs a check against all .kt files to ensure that they are formatted according to ktlint."
             configurePluginTask(target, extension, ktLintConfig, kotlinSourceDirectories)
         }
@@ -321,8 +312,8 @@ open class KtlintPlugin : Plugin<Project> {
         sourceSetName: String,
         ktLintConfig: Configuration,
         kotlinSourceDirectories: Iterable<*>
-    ): Task {
-        return target.taskHelper<KtlintCheckTask>(sourceSetName.sourceSetCheckTaskName()) {
+    ): TaskProvider<KtlintCheckTask> {
+        return target.registerTask(sourceSetName.sourceSetCheckTaskName()) {
             description = "Runs a check against all .kt files to ensure that they are formatted according to ktlint."
             configurePluginTask(target, extension, ktLintConfig, kotlinSourceDirectories)
         }
@@ -354,43 +345,32 @@ open class KtlintPlugin : Plugin<Project> {
         reporters.set(extension.reporters)
     }
 
-    private fun Project.getMetaKtlintCheckTask(): Task = this.tasks.findByName(CHECK_PARENT_TASK_NAME)
-        ?: this.task(CHECK_PARENT_TASK_NAME).apply {
-            group = VERIFICATION_GROUP
-            description = "Runs ktlint on all kotlin sources in this project."
-        }
-
-    private fun Project.getAndroidVariantMetaKtlintCheckTask(
+    private fun Project.createAndroidVariantMetaKtlintCheckTask(
         variantName: String,
         multiplatformTargetName: String? = null
-    ): Task {
-        val taskName = "ktlint${variantName.capitalize()}${multiplatformTargetName?.capitalize() ?: ""}Check"
-        return tasks.maybeCreate(taskName).apply {
-            group = VERIFICATION_GROUP
-            description = "Runs ktlint on all kotlin sources for android $variantName variant in this project."
-        }
+    ): TaskProvider<Task> = registerTask(variantName.androidVariantMetaCheckTaskName(multiplatformTargetName)) {
+        group = VERIFICATION_GROUP
+        description = "Runs ktlint on all kotlin sources for android $variantName variant in this project."
     }
 
-    private fun Project.getMetaKtlintFormatTask(): Task = this.tasks.findByName(FORMAT_PARENT_TASK_NAME)
-        ?: this.task(FORMAT_PARENT_TASK_NAME).apply {
-            group = FORMATTING_GROUP
-            description = "Runs the ktlint formatter on all kotlin sources in this project."
-        }
-
-    private fun Project.getAndroidVariantMetaKtlintFormatTask(
+    private fun Project.createAndroidVariantMetaKtlintFormatTask(
         variantName: String,
         multiplatformTargetName: String? = null
-    ): Task {
-        val taskName = "ktlint${variantName.capitalize()}${multiplatformTargetName?.capitalize() ?: ""}Format"
-        return tasks.maybeCreate(taskName).apply {
-            group = FORMATTING_GROUP
-            description = "Runs ktlint formatter on all kotlin sources for android $variantName" +
-                " variant in this project."
-        }
+    ): TaskProvider<Task> = registerTask(variantName.androidVariantMetaFormatTaskName(multiplatformTargetName)) {
+        group = FORMATTING_GROUP
+        description = "Runs ktlint formatter on all kotlin sources for android $variantName" +
+            " variant in this project."
     }
 
-    private fun setCheckTaskDependsOnKtlintCheckTask(project: Project, ktlintCheck: Task) {
-        project.tasks.findByName("check")?.dependsOn(ktlintCheck)
+    private fun setCheckTaskDependsOnKtlintCheckTask(
+        project: Project,
+        ktlintCheck: TaskProvider<KtlintCheckTask>
+    ) {
+        project.plugins.withType(LifecycleBasePlugin::class.java) {
+            project.tasks.named(LifecycleBasePlugin.CHECK_TASK_NAME).configure { task ->
+                task.dependsOn(ktlintCheck)
+            }
+        }
     }
 
     /*
@@ -406,4 +386,24 @@ open class KtlintPlugin : Plugin<Project> {
     private inline fun <reified T> Convention.getPluginHelper() = getPlugin(T::class.java)
 
     private fun Project.isConsolePlain() = gradle.startParameter.consoleOutput == ConsoleOutput.Plain
+
+    private class PluginHolder(
+        val target: Project
+    ) {
+        val extension = target.plugins.apply(KtlintBasePlugin::class.java).extension
+
+        val metaKtlintCheckTask: TaskProvider<Task> by lazy {
+            target.registerTask<Task>(CHECK_PARENT_TASK_NAME) {
+                group = VERIFICATION_GROUP
+                description = "Runs ktlint on all kotlin sources in this project."
+            }
+        }
+
+        val metaKtlintFormatTask: TaskProvider<Task> by lazy {
+            target.registerTask<Task>(FORMAT_PARENT_TASK_NAME) {
+                group = FORMATTING_GROUP
+                description = "Runs the ktlint formatter on all kotlin sources in this project."
+            }
+        }
+    }
 }